GeoIP

IP Geolocalization using the Geolite2 Database

Installation

The package is registered in the General registry and so can be installed at the REPL with

julia> using Pkg
julia> Pkg.add("GeoIP")

Usage

Data files

You can use MaxMind geolite2 csv files downloaded from the site. Due to the MaxMind policy, GeoLite.jl does not distribute GeoLite2 files and does not provide download utilities. For automated download it is recommended to use MaxMind GeoIP Update program. For proper functioning of GeoIP.jl you need to download GeoLite2 City datafile, usually it should have a name like GeoLite2-City-CSV_20191224.zip.

Files processing and loading provided with load() call. Directory where data is located should be located either in ENV["GEOIP_DATADIR"] or it can be passed as an argument to load function. Zip file location can be passed as an argument or it can be stored in ENV["GEOIP_ZIPFILE"]. For example

using GeoIP

geodata = load(zipfile = "GeoLite2-City-CSV_20191224.zip", datadir = "/data")

If ENV["GEOIP_DATADIR"] is set to "/data" and ENV["GEOIP_ZIPFILE"] is set to "GeoLite2-City-CSV_20191224.zip" then it is equivalent to

using GeoIP

geodata = load()

Example

You can get the ip data with the geolocate function or by using []

using GeoIP

geodata = load(zipfile = "GeoLite2-City-CSV_20191224.zip")
geolocate(geodata, ip"1.2.3.4")        # returns dictionary with all relevant information

# Equivalent to
geodata[ip"1.2.3.4"]

# Equivalent, but slower version
geodata["1.2.3.4"]

geolocate form is useful for broadcasting

geolocate.(geodata, [ip"1.2.3.4", ip"8.8.8.8"])  # returns vector of geo data.

Acknowledgements

This product uses, but not include, GeoLite2 data created by MaxMind, available from http://www.maxmind.com.

GeoIP.geolocate

GeoIP.load

GeoIP.geolocate — Method

geolocate(geodata::GeoIP.DB, ip)

Returns geolocation and other information determined in geodata by ip.

GeoIP.load — Method

load(; datadir, zipfile, blockcsvgz, citycsvgz)

Load GeoIP database from compressed CSV file or files. If zipfile argument is provided then load tries to load data from that file, otherwise it will try to load data from blockcsvgz and citycsvgz. By default blockcsvgz equals to "GeoLite2-City-Blocks-IPv4.csv.gz" and citycsvgz equals to "GeoLite2-City-Locations-en.csv.gz". datadir defines where data files are located and can be either set as an argument or read from the ENV variable GEOIP_DATADIR. In the same way if ENV variable GEOIP_ZIPFILE is set, then it is used for determining zipfile argument.